mgr/dashboard: Fix doc endpoints
UiApi endpoints were broken because the same baseurl was used
for both UiApi and Api endpoints, but the paths dict only generated
keys based off the Api baseurl.
Rather than truncate the base_url, it is left as-is to avoid extra
complexity since  _gen_paths uses the same base_url for all endpoints.

* Add typehints for dicts to fix failing mypy tests
* Remove now unused base_url variable from _gen_paths and test_docs.py

Fixes: https://tracker.ceph.com/issues/45957
Signed-off-by: Fabrizio D'Angelo <fdangelo@redhat.com>